Move matchers to separate module
